refactor: Move to pytest_plugin
