+++ pyproject.toml
@@ -128,6 +128,9 @@ lint = [
   "types-docutils",
 ]
 
+[tool.poetry.plugins.pytest11]
+libvcs = "libvcs.pytest_plugin"
+
 [tool.mypy]
 strict = true
